Rewrite crate using the native-tls crate
The purpose of this rewrite is to reuse all of awesome logic in the native-tls
crate rather than rewriting it. The native-tls crate already naturally supports
async I/O everywhere so let's just take advantage of it! As a result, this crate
is now stripped of all #[cfg] code essentially and now just provides two
extension traits for `TlsConnector` and `TlsAcceptor`. The extension traits
provide a `connect_async` and `accept_async` function, respectively, which
handles retry logic with mid-handshake streams automatically.

Currently this rewrites drops two current features of this crate: forcing usage
of OpenSSL and a rustls backend. I hope to reenable both of these features
through support in the native-tls crate shortly, however.